move `project` fixture to root `conftest`
